==> R/03_interpolation_funcs.R <==
################################################################################
### Script: Interpolation functions
### Author: josehcms
### Last updated: 2024-11-30
################################################################################

### Linear interpolation #------------------------------------------------------
interplin = function( x, x1, x2, y1, y2 ){
  ( ( y2 - y1 ) / ( x2 - x1 ) ) * ( x - x1 )  + y1
}

################################################################################

### Exponential interpolation
interpexp = function( x, x1, x2, y1, y2 ){
  r = log( y2 / y1 ) / ( x2 - x1 )
  y1 * exp( r * ( x - x1 ) )
}
################################################################################
